+++ src/agents/user_agent.py
@@ -1,33 +1,38 @@
+import calendar
 import json
+import time
+from dataclasses import dataclass
 from typing import List
+
 from spade.agent import Agent
 from spade.behaviour import CyclicBehaviour
 from spade.message import Message
 from spade.template import Template
 
-from dataclasses import dataclass
-import calendar
-import time
-
 from constants import DEFAULT_HOST, MESSAGE_TIMEOUT
+from logger import logger
 from messages.check_offers import CheckOffers
-from messages.consolidated_offers import ConsolidatedOffers
+from messages.consolidated_offers import ConsolidatedOffers, Offer
 from messages.modify_reservation import ModifyReservation
 from messages.reservation_request import RequestReservation
 from messages.response import ReservationResponse
 
 
 class User(Agent):
-    def __init__(self, jid, password):
+    def __init__(self, jid, password, coordinators_jids=[]):
         super().__init__(jid, password)
-        self.coordinators_jids = None
+        self.coordinators_jids = coordinators_jids
         self.pending_reservation = None
         # dictionary of active reservations {reservation_id: Reservation}
         self.active_reservations = {}
 
     def _validate_timeslot(self, time_start, time_stop):
-        current_time = calendar.timegm(time.gmtime())  # Get current timestamp in UTC timezone
-        if not (time_start < time_stop and time_start > current_time):
+        # current_time = calendar.timegm(time.gmtime())  # Get current timestamp in UTC timezone
+        # if not (time_start < time_stop and time_start > current_time):
+        #     raise ValueError("Invalid timeslot")
+        if not (time_start < time_stop):
+            raise ValueError("Invalid timeslot")
+        if time_start < 0 or time_stop > 24:
             raise ValueError("Invalid timeslot")
 
     def _validate_location(self, x, y):
@@ -36,14 +41,14 @@ class User(Agent):
 
     def _prepare_consolidated_offers_template(self):
         template = Template()
-        template.to = f"{self.jid}@{DEFAULT_HOST}"
+        template.to = f"{self.jid}"
         template.set_metadata("performative", "inform")
         template.set_metadata("action", "consolidated-offers")
         return template
 
     def _prepare_reservation_response_template(self):
         template = Template()
-        template.to = f"{self.jid}@{DEFAULT_HOST}"
+        template.to = f"{self.jid}"
         template.set_metadata("performative", "inform")
         template.set_metadata("action", "reservation-response")
         return template
@@ -64,10 +69,12 @@ class User(Agent):
         for coordinator_jid in self.coordinators_jids:
             to_send = Message(
                 to=coordinator_jid,
-                body=json.dumps(check_offers_msg.dict()),
+                body=check_offers_msg.model_dump_json(),
                 metadata={"performative": "query-ref", "action": "check-offers"},
             )
-            await self.send(to_send)
+            aioxmpp_msg = to_send.prepare()
+            await self.client.send(aioxmpp_msg)
+        logger.info(f"Check offers message sent to coordinators: {self.coordinators_jids}")
 
     async def request_reservation_modification(self, reservation_id, time_start, time_stop):
         self._validate_timeslot(time_start, time_stop)
@@ -76,31 +83,33 @@ class User(Agent):
             self.pending_reservation.time_start = time_start
             self.pending_reservation.time_stop = time_stop
             modify_reservation_msg = ModifyReservation(
-                reservation_id=reservation_id, time_start=time_start, time_stop=time_stop, user_id=self.jid
+                reservation_id=reservation_id, time_start=time_start, time_stop=time_stop, user_id=str(self.jid)
             )
+            # TODO critical: message.to should be correct regional coordinator
             to_send = Message(
                 to=f"{self.jid}@{DEFAULT_HOST}",
-                body=json.dumps(modify_reservation_msg.dict()),
+                body=modify_reservation_msg.model_dump_json(),
                 metadata={"performative": "request", "action": "modify-reservation"},
             )
             await self.send(to_send)
+            logger.info(f"Modify reservation message sent: {modify_reservation_msg.model_dump_json()}")
         else:
             raise Exception("Reservation not found")
 
-    def choose_parking_offer(self, offers: List[dict]):
+    def choose_parking_offer(self, offers: List[Offer]):
         # TODO implement logic of choosing an offer
-        return offers[0]["parking_id"] if offers else None
+        return offers[0].parking_id if offers else None
 
     def save_reservation(self, reservation_response: ReservationResponse):
-        if not self.agent.pending_reservation:
+        if not self.pending_reservation:
             raise Exception("No pending reservation")
         # If reservation id is set -> there was a modification request
-        if self.agent.pending_reservation.id:
+        if self.pending_reservation.id:
             # The old reservation will be replaced by the new one with a new id
-            self.agent.active_reservations.pop(self.agent.pending_reservation.id)
-        self.agent.pending_reservation.id = reservation_response.reservation_id
-        self.agent.active_reservations[reservation_response.reservation_id] = self.agent.pending_reservation
-        self.agent.pending_reservation = None
+            self.active_reservations.pop(self.pending_reservation.id)
+        self.pending_reservation.id = reservation_response.reservation_id
+        self.active_reservations[reservation_response.reservation_id] = self.pending_reservation
+        self.pending_reservation = None
 
     class MakeReservation(CyclicBehaviour):
         """Behaviour for making a reservation
@@ -110,20 +119,29 @@ class User(Agent):
         async def run(self):
             msg = await self.receive(timeout=MESSAGE_TIMEOUT)
             if msg:
-                consolidated_offers = ConsolidatedOffers(**json.loads(msg.body))
+                consolidated_offers = ConsolidatedOffers.model_validate_json(msg.body)
+                logger.info(
+                    f"{str(self.agent.jid)}: Received ConsolidatedOffers from {msg.sender}: {consolidated_offers.model_dump_json()}"
+                )
                 chosen_parking_id = self.agent.choose_parking_offer(consolidated_offers.offers)
                 if self.agent.pending_reservation:
                     self.agent.pending_reservation.parking_id = chosen_parking_id
                     res = self.agent.pending_reservation
                     reservation_request = RequestReservation(
-                        res.time_start, res.time_stop, chosen_parking_id, self.agent.jid
+                        time_start=res.time_start,
+                        time_stop=res.time_stop,
+                        parking_id=chosen_parking_id,
+                        user_id=str(self.agent.jid),
                     )
                     to_send = Message(
-                        to=consolidated_offers.sender,
-                        body=json.dumps(reservation_request.dict()),
+                        to=str(msg.sender),
+                        body=reservation_request.model_dump_json(),
                         metadata={"performative": "request", "action": "make-reservation"},
                     )
                     await self.send(to_send)
+                    logger.info(
+                        f"{str(self.agent.jid)}: Reservation request to {str(msg.sender)} sent: {reservation_request.model_dump_json()}"
+                    )
 
     class AwaitReservationConfirmation(CyclicBehaviour):
         """Behaviour for waiting for reservation confirmation"""
@@ -131,12 +149,18 @@ class User(Agent):
         async def run(self):
             msg = await self.receive(timeout=MESSAGE_TIMEOUT)
             if msg:
-                reservation_response = ReservationResponse(**json.loads(msg.body))
+                reservation_response = ReservationResponse.model_validate_json(msg.body)
+                logger.info(
+                    f"{str(self.agent.jid)}: Received ReservationResponse: {reservation_response.model_dump_json()}"
+                )
                 if reservation_response.success:
                     self.agent.save_reservation(reservation_response)
+                    logger.info(
+                        f"{str(self.agent.jid)}: Reservation successful. Reservation id: {reservation_response.reservation_id}"
+                    )
                 else:
                     # TODO Implement reservation failue handling
-                    print("Reservation failed.")
+                    logger.error(f"{str(self.agent.jid)}: Reservation failed.")
 
     async def setup(self):