+++ src/messages/check_parking.py
@@ -1,11 +1,8 @@
-from dataclasses import dataclass
-
-from messages.base import BaseMessage
+from pydantic import BaseModel
 
 
-@dataclass
-class CheckParking(BaseMessage):
+class CheckParking(BaseModel):
     """Message to check parking availability in specified time slot"""
 
-    time_start: int  # assume timestamp
-    time_stop: int  # assume timestamp
+    time_start: int
+    time_stop: int