+++ src/messages/modify_reservation.py
@@ -1,14 +1,11 @@
-from dataclasses import dataclass
-
-from messages.base import BaseMessage
+from pydantic import BaseModel
 
 
-@dataclass
-class ModifyReservation(BaseMessage):
+class ModifyReservation(BaseModel):
     """Message from RegionalCoordinator to Parking
     to modify a reservation to new time slot"""
 
-    time_start: int  # assume timestamp
-    time_stop: int  # assume timestamp
+    time_start: int
+    time_stop: int
     reservation_id: str
     user_id: str